requirements.txt
streamlit
streamlit-authenticator
pandas
pyyaml
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
